ex. 4.6-4.7 & ex. 4.21-4.22

+++ Osa4/bloglist/controllers/blogs.js
@@ -2,13 +2,18 @@ const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
 const jwt = require("jsonwebtoken");
+const { userExtractor } = require("../utils/middleware");
 
 blogsRouter.get("/", async (request, response) => {
-  const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
+  const blogs = await Blog.find({}).populate("user", {
+    username: 1,
+    name: 1,
+    id: 1,
+  });
   response.json(blogs);
 });
 
-blogsRouter.post("/", async (request, response) => {
+blogsRouter.post("/", userExtractor, async (request, response) => {
   const body = request.body;
   const decodedToken = jwt.verify(request.token, process.env.SECRET);
   if (!decodedToken.id) {
@@ -24,12 +29,21 @@ blogsRouter.post("/", async (request, response) => {
     user: user._id,
   });
   const savedBlog = await blog.save();
-  user.blogs = user.blogs.concat(savedBlog._id);
-  await user.save();
+  request.user.blogs = request.user.blogs.concat(savedBlog._id);
+  await request.user.save();
   response.status(201).json(savedBlog);
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
+blogsRouter.delete("/:id", userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id);
+  if (!blog) {
+    return response.status(400).json({ error: "invalid token " });
+  }
+  if (blog.user.toString() !== request.user.id.toString()) {
+    return response
+      .status(400)
+      .json({ error: "users can only delete own posts" });
+  }
   await Blog.findByIdAndDelete(request.params.id);
   response.status(204).end();
 });
